+++ Changelog.md
@@ -1,4 +1,13 @@
-### New Sample Files Available
+### 1.1.1 - External Service Sample
+In version 1.1.1 there were added sample files to consume an external odata service exposed as entity of the internal catalog service.
+The feature can be used with:
+```sh
+npx cap-utils --add ext
+```
+
+### 1.0.7 - New Sample Files Available
+From version 1.0.11 you can modify your package.json to add a cds profiling section.
+
 In version 1.0.7 sample files for xsuaa,html5-repo,hana,mta based deployment were added:
 - valid-service.cds were changed to add scope protection on srv
 - test.http were modified to add Basic authentication for local testing
@@ -6,4 +15,5 @@ In version 1.0.7 sample files for xsuaa,html5-repo,hana,mta based deployment wer
 - xs-security-cc.json : it will add a client credential user with authorizations to consume srv
 - mta.yaml : enable the app for a Cloud Foundry deployment. 
 
-From version 1.0.11 you can modify your package.json to add a cds profiling section.
+
+